chore(ios): split settings into separate view files
